specification: Add a specification of OARS 1.1
Add a prose specification of the semantics of OARS 1.0 and 1.1, and
RelaxNG Compact schemas of the structure of both.

The semantics are based on the behaviour of the generator, which was
previously the canonical standard.

Add a `meson.build` with tests to validate the schemas, although this is
only present as documentation for the time being, since there’s no build
system for the module as a whole.

Signed-off-by: Philip Withnall <withnall@endlessm.com>